Add a small delay before reading the button state initially.
This is needed since on unit #002 the lack of delay was causing the
initial reading of the button state (to determine whether to enter
uSv/h or counter mode) to behave randomly. I guess some stray
capacitance in the MCU internally makes it unsafe to read the button
state immediately after turning the WPU on - the input pin buffer is
probably still catching up by the time we read it.

1 ms is probably a huge overkill, but noone would notice it anyway.