Restore piezo tone. It shall be 3125 Hz regardless of F_CPU.
The 6 MHz crystal I'm testing with now made the GM clicks sound at
quite a lower pitch. The code now computes the correct PWM frequency,
regardless of the speed of the crystal (however, the crystal speed
should be an integer number in MHz. As we're only using 6MHz and 8MHz
crystals currently, that shouldn't be a problem).