Fix a bug in ADC sampling.
On Atmels, it's very important to use the correct order of access to
the ADCL and ADCH registers. Using the ADCW gives us the 10-bit value
directly, and removes the burden of thinking how the compiler will
order reads to ADCL/ADCH.

The bug was that due to incorrect order, ADCL/ADCH registers were
"in lock" for eternity, and only the first ever ADC sample was being
returned all the time.