Implement long logging.
Long logging is the idea of logging radiation over time in the
EEPROM of the ATmega. Since the space there is limited, we
support two types of subsampling. The first one is that
the GM counts detected are initially binned in 30 seconds
intervals, but if the count of intervals exceeds 240 (2 hours),
they are re-binned into 60 second intervals (and can be
further re-binned ad infinitum).

The second one is that each bin is a 16-bit value and for
high radiation values, or long logging times, these might
overflow. We deal with that by storing a separate scaling
coefficient, termed 'scaling'. In effect, values in the log
are later interpreted as "value << scaling".

We also introduce a PC link to the device, which enables
querying its status and fetching the stored log.

For more details, consult the comment section in pc_link.c